+++ .eslintrc.js
@@ -1,13 +1,15 @@
 // https://eslint.org/docs/user-guide/configuring
 
 module.exports = {
+  root: true,
+  parser: "@typescript-eslint/parser",
   env: {
     browser: true,
     node: true,
-    es6: true
+    es6: true,
+    jest: true,
   },
-
-  globals: {
-    page: 'readonly'
-  }
+  // https://github.com/standard/standard/blob/master/docs/RULES-en.md
+  extends: ["standard", "prettier"],
+  plugins: ["@typescript-eslint"],
 }