+++ tailwind.config.js
@@ -1,6 +1,3 @@
-module.exports = {
-  content: [
-    "./index.html",
-    "./src/**/*.{js,ts,jsx,tsx}",
-  ]
+export default {
+  content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
 }