+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -6,9 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.0.0] - 2024-03-23
+
 ### Added
 
 - Dispatch `rails-nested-form:add` and `rails-nested-form:remove` events.
+
+### Chore
+
+- Renaming the component from `stimulus-rails-nested-form` to `@stimulus-components/rails-nested-form`
+- Upgrading dependencies
+- Exporting Typescript Types
+- Updating demo UI
+- Add [Stimulus LSP](https://github.com/marcoroth/stimulus-lsp) compatibility
 
 ## [4.1.0] - 2022-12-23